Fix #15 : Added I2C Functions
