+++ PSLab_Original/functions.h
@@ -103,15 +103,15 @@ unsigned int dest[_FLASH_ROW * 8];
 int ulsb,umsb; //DAC_OFFSETS[4],
 
 /*-----SPI VARIABLES-------*/
-BYTE location, value,ADC_MODE=NOT_READY,SPI_MODE=NOT_READY,MULTIFUNC_PORT = NOT_READY,DMA_MODE=NOT_READY,ADC_STREAMING=false;
+BYTE location, value,ADC_MODE=NOT_READY,SPI_MODE=NOT_READY,DMA_MODE=NOT_READY,ADC_STREAMING=false;
 BYTE SPI_PPRE=0,SPI_SPRE=2,SPI_CKE=1,SPI_CKP=0,SPI_SMP=1;
 
 /*------UART VARIABLES-----*/
-unsigned int I2C_BRGVAL = 0x272,TCD = 1000;
+unsigned int TCD = 1000;
 
 /*------LOGIC ANALYZER VARIABLES-----*/
 BYTE INITIAL_DIGITAL_STATES=0,INITIAL_DIGITAL_STATES_ERR=0,DIGITAL_TRIGGER_CHANNEL=32,DIGITAL_TRIGGER_STATE=0,b1,b2,COMPARATOR_CONFIG=7|(3<<4),conversion_done = 1,I2CConvDone = 1;
-unsigned int i, lsb, msb, blk[8], adval,tmp_int1,tmp_int2,tmp_int3,tmp_int4,tmp_int5,tmp_int6;
+unsigned int i, lsb, msb, blk[8], adval,tmp_int2,tmp_int3,tmp_int4,tmp_int5,tmp_int6;
 
 unsigned int LAFinished = 1, LASamples;
 unsigned int samples_to_fetch = BUFFER_SIZE, I2CTotalSamples = BUFFER_SIZE;
@@ -194,7 +194,6 @@ void start_3chan_LA(unsigned int ,unsigned int ,BYTE );
 
 void start_4chan_LA(unsigned int,unsigned int,BYTE);
 void disable_input_capture();
-void setMultiFuncPortMode(BYTE);
 
 void init(void);
 void setFlashPointer(BYTE);
@@ -242,20 +241,6 @@ void setSensorChannel(char);
 void read_all_from_flash(_prog_addressT pointer);
 void load_to_flash(_prog_addressT pointer, BYTE location, unsigned int * blk);
 void read_flash(_prog_addressT pointer, BYTE location);
-
-void initI2C(void);
-
-
-void I2CStart();
-void I2CStop();
-void I2CRestart();
-void I2CAck();
-void I2CNak();
-void I2CWait();
-void I2CSend(BYTE dat);
-BYTE I2CRead(BYTE ack);
-
-
 
 /*Command set for the NRFL01+ radio*/
 void nRF_Setup();
